+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.17.11](https://github.com/projen/projen/compare/v0.17.10...v0.17.11) (2021-03-17)
+
+
+### Features
+
+* **task:** add TaskStepOptions to task reset ([#619](https://github.com/projen/projen/issues/619)) ([996b0c2](https://github.com/projen/projen/commit/996b0c24b8ff46f2e52f65e9e64b8b1d9f484934))
+
+### [0.17.10](https://github.com/projen/projen/compare/v0.17.9...v0.17.10) (2021-03-17)
+
+
+### Features
+
+* **node:** mutable pull request builds ([#624](https://github.com/projen/projen/issues/624)) ([757f8b1](https://github.com/projen/projen/commit/757f8b12a3a93aa83608a7dd18f171738088665c)), closes [#602](https://github.com/projen/projen/issues/602) [#604](https://github.com/projen/projen/issues/604)
+
+### [0.17.9](https://github.com/projen/projen/compare/v0.17.8...v0.17.9) (2021-03-17)
+
+### [0.17.8](https://github.com/projen/projen/compare/v0.17.7...v0.17.8) (2021-03-17)
+
+### [0.17.7](https://github.com/projen/projen/compare/v0.17.6...v0.17.7) (2021-03-17)
+
+### [0.17.6](https://github.com/projen/projen/compare/v0.17.5...v0.17.6) (2021-03-16)
+
+### [0.17.5](https://github.com/projen/projen/compare/v0.17.4...v0.17.5) (2021-03-16)
+
+### [0.17.4](https://github.com/projen/projen/compare/v0.17.3...v0.17.4) (2021-03-15)
+
+
+### Bug Fixes
+
+* **node-project:** fix projenVersion type to correctly use a string ([#608](https://github.com/projen/projen/issues/608)) ([dbc3c9c](https://github.com/projen/projen/commit/dbc3c9ca84ba6e0d8990d2b326bd00001787d5ea))
+
+### [0.17.3](https://github.com/projen/projen/compare/v0.17.2...v0.17.3) (2021-03-15)
+
+### [0.17.2](https://github.com/projen/projen/compare/v0.17.1...v0.17.2) (2021-03-15)
+
 ### [0.17.1](https://github.com/projen/projen/compare/v0.17.0...v0.17.1) (2021-03-11)